Rename ulaMessage.ts to ula-message.ts
