+++ src/index.ts
@@ -18,7 +18,7 @@ export { EventHandler } from './event-handler'
 
 export { HttpHandler } from './http-handler'
 
-export { UlaMessage, Message } from './model/ulaMessage'
+export { UlaMessage, Message } from './model/ula-message'
 
 export { IAttestation, Attestation } from './model/attestation'
